backend/requirements.txt
agentscope>=1.0.0
fastapi>=0.115.0
starlette>=0.37.0
uvicorn[standard]>=0.30.0
python-multipart==0.0.9
pydantic>=2.6.1
pydantic-settings>=2.5.0
openai==1.12.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pdfplumber==0.10.3
httpx>=0.27.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
pytesseract>=0.3.10
python-dotenv>=1.0.0